Rename `new_video_capture` to `video_capture`
