+++ pupil_src/capture/eye.py
@@ -101,11 +101,11 @@ def eye(timebase, is_alive_flag, ipc_pub_url, ipc_sub_url,ipc_push_url, user_dir
         from file_methods import Persistent_Dict
         from version_utils import VersionFormat
         from methods import normalize, denormalize, Roi, timer
-        from video_capture import FileCaptureError, EndofVideoFileError, CameraCaptureError
         from av_writer import JPEG_Writer,AV_Writer
 
-        from new_video_capture.source import InitialisationError, Fake_Source
-        from new_video_capture import source_classes, manager_classes
+        from video_capture.source import InitialisationError, Fake_Source
+        from video_capture.source.file_source import EndofVideoFileError
+        from video_capture import source_classes, manager_classes
         source_by_name = {src.class_name():src for src in source_classes}
 
         # Pupil detectors
