+++ pupil_src/capture/world.py
@@ -82,9 +82,9 @@ def world(timebase,eyes_are_alive,ipc_pub_url,ipc_sub_url,ipc_push_url,user_dir,
     from file_methods import Persistent_Dict
     from methods import normalize, denormalize, delta_t, get_system_info
 
-    from new_video_capture.source import InitialisationError, Fake_Source
-    from new_video_capture.source.file_source import EndofVideoFileError
-    from new_video_capture import source_classes, manager_classes
+    from video_capture.source import InitialisationError, Fake_Source
+    from video_capture.source.file_source import EndofVideoFileError
+    from video_capture import source_classes, manager_classes
     source_by_name = {src.class_name():src for src in source_classes}
 
     from version_utils import VersionFormat
