+++ pupil_src/shared_modules/video_capture/manager/__init__.py
