+++ pupil_src/shared_modules/video_capture/manager/fake_manager.py
