+++ pupil_src/shared_modules/video_capture/manager/ndsi_manager.py
