+++ pupil_src/shared_modules/video_capture/manager/uvc_manager.py
